bot/requirements/base.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiohttp==3.8.4
aiosignal==1.3.1
APScheduler==3.10.1
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
emoji==2.2.0
frozenlist==1.3.3
gunicorn==20.1.0
h11==0.14.0
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
prettytable==3.8.0
pydantic==1.10.9
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.15
tinkoff==0.1.1
tinkoff-investments==0.2.0b58
typing_extensions==4.6.3
tzlocal==5.0.1
uvicorn==0.22.0
yarl==1.9.2
